Add test setup
