+++ package.json
@@ -21,7 +21,22 @@
     "homepage": "https://github.com/GoodbyePlanet/share-it-with-me#readme",
     "scripts": {
         "build": "rm -rf dist && tsc",
-        "start": "dotenv -- nodemon -e ts,graphql -x ts-node src/index.ts"
+        "start": "dotenv -- nodemon -e ts,graphql -x ts-node src/index.ts",
+        "test": "NODE_ENV=test dotenv jest"
+    },
+    "jest": {
+        "transform": {
+            "^.+\\.tsx?$": "ts-jest"
+        },
+        "testRegex": "(/__tests__/.*|(\\.|/)(test|spec))\\.(jsx?|tsx?)$",
+        "moduleFileExtensions": [
+            "ts",
+            "tsx",
+            "js",
+            "jsx",
+            "json",
+            "node"
+        ]
     },
     "devDependencies": {
         "@types/bcryptjs": "^2.4.2",
@@ -29,8 +44,11 @@
         "@types/node": "^14.14.19",
         "@types/ws": "^7.4.0",
         "dotenv-cli": "^4.0.0",
+        "graphql-request": "^3.4.0",
+        "jest": "^26.6.3",
         "nodemon": "^2.0.6",
         "prisma": "^1.34.10",
+        "ts-jest": "^26.5.0",
         "ts-node": "^9.1.1",
         "typescript": "^4.1.3"
     },
